+++ lfsr/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] - ReleaseDate
+### Added
+* Direct lookup tables example
+### Changed
+* Renamed `lfsr_lookup!` to `searching_lfsr_lookup!`
 
 ## [0.2.0] - 2019-06-07
 ### Initial release